version final C++

+++ cpp/include/IHeuristic.h
@@ -3,12 +3,14 @@
 
 #include "Node.h"
 
+
+
 class IHeuristic
 {
 public:
 	IHeuristic() {};
 	virtual ~IHeuristic(void) {};
-	virtual int getH(Node * beforeState, Node* currentState) = 0;
+	virtual double getH(Node * beforeState, Node* currentState) = 0;
 	virtual void setSolution(Node *solution) = 0;
 };
